16 Crew-AI-Crash-course-main/requirements.txt
crewai
crewai_tools
load_dotenv
langchain-huggingface
ollama
langchain
langchain-community
sentence-transformers
